Refactor fastq file concatenation logic
